Refactor vehicle behavior and upgrade Rider package
In this commit, the car behavior script (CarMovementBehaviour.cs) was renamed to VehicleBehavoiur.cs, and several features were updated. The class was refactored to improve readability and the rigidbody and collider variable names were prefixed with an underscore for consistency. The `Car` object has been replaced with `VehicleSettings` to make the code more general. Also, the vehicle now has its own movement, separate from the player. Additionally, a `_rigidbody` Unity class property was added to replace the previously direct-access `rigidbody` class property.

A number of new features were provided- including a better vehicle direction control, improved exit position maintenance, and a more refined wheel rotation system.
The commit introduced an ideal environment for the vehicle-wheel movements, which improves the VR as well as the real-world understanding of the game.

Overall, the commit was majorly to improve the vehicle movements, player-related variables, and code readability.